DB fixes

+++ modules/utils.py
@@ -17,32 +17,33 @@ from modules import globals, db, errors, xp
 
 # Get database
 async def get_db():
-    if globals.DB_HOST_TYPE == "github":
-        async with globals.http.get(f'https://gist.githubusercontent.com/{globals.GITHUB_GIST_USER}/{globals.GITHUB_GIST_ID}/raw',
-                                    headers={
-                                        'Authorization': f'Token {globals.GITHUB_GIST_TOKEN}'
-                                    }) as req:
-            db_data = await req.text()
-    elif globals.DB_HOST_TYPE == "writeas":
-        async with globals.http.post('https://write.as/api/auth/login',
-                                     headers={
-                                         'Content-Type': 'application/json'
-                                     },
-                                     data=json.dumps({
-                                         "alias": globals.WRITE_AS_USER,
-                                         "pass": globals.WRITE_AS_PASS
-                                     })) as req:
-            globals.write_as_token = (await req.json())["data"]["access_token"]
-        async with globals.http.get(f"https://write.as/{globals.WRITE_AS_USER}/{globals.WRITE_AS_POST_ID}.txt") as req:
-            db_data = await req.text()
-    else:
-        raise Exception("No valid DB type specified!")
-    decoded = base64.b85decode(db_data.encode("utf-8"))
-    decompressed = zlib.decompress(decoded)
-    async with aiofiles.open('db.sqlite3', 'wb') as f:
-        await f.write(decompressed)
+    if not os.path.exists("db.sqlite3"):
+        if globals.DB_HOST_TYPE == "github":
+            async with globals.http.get(f'https://gist.githubusercontent.com/{globals.GITHUB_GIST_USER}/{globals.GITHUB_GIST_ID}/raw',
+                                        headers={
+                                            'Authorization': f'Token {globals.GITHUB_GIST_TOKEN}'
+                                        }) as req:
+                db_data = await req.text()
+        elif globals.DB_HOST_TYPE == "writeas":
+            async with globals.http.post('https://write.as/api/auth/login',
+                                         headers={
+                                             'Content-Type': 'application/json'
+                                         },
+                                         data=json.dumps({
+                                             "alias": globals.WRITE_AS_USER,
+                                             "pass": globals.WRITE_AS_PASS
+                                         })) as req:
+                globals.write_as_token = (await req.json())["data"]["access_token"]
+            async with globals.http.get(f"https://write.as/{globals.WRITE_AS_USER}/{globals.WRITE_AS_POST_ID}.txt") as req:
+                db_data = await req.text()
+        else:
+            raise Exception("No valid DB type specified!")
+        decoded = base64.b85decode(db_data.encode("utf-8"))
+        decompressed = zlib.decompress(decoded)
+        async with aiofiles.open('db.sqlite3', 'wb') as f:
+            await f.write(decompressed)
     await db.init_db()
-    print("Fetched DB!")
+    print("Initialized DB!")
 
 
 # Save database
@@ -92,6 +93,14 @@ async def save_db():
 
 # Restart the bot, dyno restart on heroku
 async def restart():
+    print("Saving DB...")
+    await db.save_to_disk()
+    admin = globals.bot.get_user(globals.ADMIN_ID)
+    if admin:
+        await admin.send(file=discord.File('db.sqlite3'))
+    await save_db()
+    await globals.db.close()
+    print("Restarting...")
     if os.environ.get("DYNO"):
         async with globals.http.delete(f'https://api.heroku.com/apps/{os.environ.get("HEROKU_APP_NAME")}/dynos/{os.environ["DYNO"]}',
                                        headers={
